app/requirements.txt
uwsgi
flask
pillow